refactor: split MechConfigurationSystem.ts and VaultInterfaces.ts
Task 4.4: MechConfigurationSystem.ts (1136 → 9 files)
- Created MechConfigTypes.ts (188 lines) - types, interfaces, enums
- Created mechConfigHelpers.ts (302 lines) - shared helpers
- Created 5 config files (49-140 lines each)
- Created mechConfigRegistry.ts (144 lines) - registry singleton
- Barrel re-export (17 lines)

Task 4.5: VaultInterfaces.ts (1269 → 7 files)
- Created 6 domain files following dependency order
- VaultCoreTypes → ImportExport → Sharing → Contact → Sync → Versioning
- Each file 98-155 lines, all under threshold
- Barrel re-export (26 lines)

All files under FILE_MODULARITY_SPEC thresholds. Zero consumer changes.
